add helper class to combine all repeated code
