+++ src/test/kotlin/usecase/ViewTaskLogsUseCaseTest.kt
@@ -2,34 +2,20 @@ package usecase
 
 import io.mockk.every
 import io.mockk.mockk
-import kotlinx.datetime.LocalDateTime
-import logic.entities.LogItem
 import logic.exception.PlanMateException.NotFoundException.TaskIDNotFoundException
 import logic.repository.LogRepository
 import logic.usecases.ViewTaskLogsUseCase
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
+import utils.taskLogs
 import java.util.UUID
 import kotlin.test.assertEquals
 
 class ViewTaskLogsUseCaseTest {
     private lateinit var repository : LogRepository
     private lateinit var useCase : ViewTaskLogsUseCase
-    private val taskLogs = listOf<LogItem>(
-        LogItem(
-            id = UUID.fromString("11111111-1111-1111-1111-111111111111"),
-            message = "This task name is updated by Hend at 12:30 29-4-2025",
-            date = LocalDateTime(2025, 4, 29, 8, 30),
-            entityId = UUID.fromString("123e4567-e89b-12d3-a456-426614174000"),
-        ),
-        LogItem(
-            id = UUID.fromString("11111551-22a1-bb11-1111-177588lkhjk1"),
-            message = "This task state is updated by Hend at 01:30 29-4-2025",
-            date = LocalDateTime(2025, 4, 29, 8, 30),
-            entityId = UUID.fromString("123e4567-e89b-12d3-a456-426614174000"),
-        )
-    )
+
 
     @BeforeEach
     fun setup(){
@@ -40,10 +26,10 @@ class ViewTaskLogsUseCaseTest {
     @Test
     fun `should return logs of task when id of task is found`(){
         val id = UUID.fromString("123e4567-e89b-12d3-a456-426614174000")
-        every { repository.viewLogsById(id) } returns taskLogs
+        every { repository.viewLogsById(id) } returns taskLogs()
 
         val returnedLogs = useCase.viewTaskLogs(id)
-        assertEquals(taskLogs, returnedLogs)
+        assertEquals(taskLogs(), returnedLogs)
     }
 
    @Test
